+++ R/scale_break_functions.R
@@ -14,7 +14,7 @@ integer_breaks <- function(n = 5, ...) {
 
 #' breaks function for including min + max with labels, and a few unlabeled ticks in between
 #' @param n number of breaks added within the min-max range
-#' @param digits number of decimal places to display 
+#' @param digits number of decimal places to display
 range_breaks <- function(n=3, digits=2) {
   fxn <- function(x) {
     if (is.null(x) || all(is.na(x) || all(is.infinite(x)))) {